# packages/synth-containers/openapi/container-contract-v1.yaml
openapi: 3.1.0
info:
  title: Synth Containers Contract
  version: 0.1.0
  description: |
    Canonical public wire contract for the standardized Synth container runtime.
    It covers metadata discovery, task discovery, blocking and async rollouts,
    rollout lifecycle polling, checkpoints, resume semantics, and control
    surfaces such as pause and terminate.
servers:
  - url: http://localhost
    description: Local development
paths:
...
  /task_info:
    get:
      operationId: getTaskInfo
      summary: Read task and capability metadata for the active runtime
      parameters:
        - name: seed
          in: query
          required: false
          schema:
            type: integer
          description: Optional seed selector for runtimes with request-aware task info.
        - name: split_group
          in: query
          required: false
          schema:
            type: string
          description: Optional split selector for catalog-backed runtimes.
        - name: family
          in: query
          required: false
          schema:
            type: string
          description: Optional task family selector.
        - name: task_instance_id
          in: query
          required: false
          schema:
            type: string
          description: Optional concrete task instance selector.
        - name: task_id
          in: query
          required: false
          schema:
            type: string
          description: Optional task selector.
      responses:
        '200':
          description: Task info payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskInfoResponse'
...
  /rollout:
    post:
      operationId: submitRollout
      summary: Start a rollout using sync or async submission mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolloutRequest'
      responses:
        '200':
          description: Rollout completed synchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
        '202':
          description: Rollout accepted asynchronously and queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutResponse'
        '422':
          description: Request validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
components:
...
  schemas:
    SubmissionMode:
      type: string
      enum: [sync, async]
...
    ResourceKind:
      type: string
      enum:
        - data
        - code
        - runtime
        - state
        - evaluation
        - tooling
        - secret
        - config
        - artifact
        - other
    ResourceRef:
      type: object
      required: [resource_id]
      properties:
        resource_id:
          type: string
        kind:
          $ref: '#/components/schemas/ResourceKind'
        name:
          type: string
        subtype:
          type: string
        uri:
          type: string
        path:
          type: string
        digest:
          type: string
        version:
          type: string
        media_type:
          type: string
        size_bytes:
          type: integer
        required:
          type: boolean
          default: true
        labels:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
      additionalProperties: true
...
    TaskInfoResponse:
      type: object
      properties:
        task:
          type: object
          additionalProperties: true
        dataset:
          type: object
          additionalProperties: true
        capabilities:
          type: object
          additionalProperties: true
        limits:
          type: object
          additionalProperties: true
        inference:
          type: object
          additionalProperties: true
        task_metadata:
          type: object
          additionalProperties: true
        environment:
          type: string
        metadata:
          type: object
          additionalProperties: true
        resource_refs:
          type: array
          items:
            $ref: '#/components/schemas/ResourceRef'
      additionalProperties: true
...
    RolloutEnvSpec:
      type: object
      properties:
        config:
          type: object
          additionalProperties: true
        seed:
          type: integer
      additionalProperties: true
    RolloutPolicySpec:
      type: object
      properties:
        config:
          type: object
          additionalProperties: true
      additionalProperties: true
    RolloutDatasetSpec:
      type: object
      properties:
        path:
          type: string
        split:
          type: string
        limit:
          type: integer
        config:
          type: object
          additionalProperties: true
      additionalProperties: true
    RolloutActorSpec:
      type: object
      required: [actor_id]
      properties:
        actor_id:
          type: string
        role:
          type: string
        config:
          type: object
          additionalProperties: true
      additionalProperties: false
    RewardInfo:
      type: object
      properties:
        outcome_reward:
          type: number
          default: 0
        event_rewards:
          type: array
          items:
            type: object
            additionalProperties: true
        score:
          type: number
        metrics:
          type: object
          additionalProperties: true
      additionalProperties: true
    ArtifactItem:
      type: object
      properties:
        artifact_id:
          type: string
        kind:
          type: string
          default: generic
        uri:
          type: string
        path:
          type: string
        media_type:
          type: string
        digest:
          type: string
        name:
          type: string
        metadata:
          type: object
          additionalProperties: true
      additionalProperties: true
    RolloutRequest:
      type: object
      properties:
        trace_correlation_id:
          type: string
        trial_id:
          type: string
        run_id:
          type: string
        mode:
          type: string
        submission_mode:
          $ref: '#/components/schemas/SubmissionMode'
        env:
          $ref: '#/components/schemas/RolloutEnvSpec'
        policy:
          $ref: '#/components/schemas/RolloutPolicySpec'
        dataset:
          $ref: '#/components/schemas/RolloutDatasetSpec'
        long_horizon:
          type: object
          additionalProperties: true
        terminator:
          type: object
          additionalProperties: true
        task_payload:
          type: object
          additionalProperties: true
        task_id:
          type: string
        task_instance_id:
          type: string
        task_metadata:
          type: object
          additionalProperties: true
        checkpoint:
          oneOf:
            - type: object
              additionalProperties: true
            - type: string
        checkpoint_id:
          type: string
        checkpoint_data_base64:
          type: string
        target_rollout_id:
          type: string
        actors:
          type: array
          items:
            $ref: '#/components/schemas/RolloutActorSpec'
        actor_ids:
          type: array
          items:
            type: string
        actor_overrides:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
      additionalProperties: false
    RolloutResponse:
      type: object
      required: [trace_correlation_id, rollout_id]
      properties:
        trace_correlation_id:
          type: string
        rollout_id:
          type: string
        trial_id:
          type: string
        status:
          type: string
          default: completed
        success_status:
          type: string
          default: success
        status_detail:
          type: string
        summary:
          type: object
          additionalProperties: true
        reward_info:
          $ref: '#/components/schemas/RewardInfo'
        trace:
          type: object
          additionalProperties: true
        turns:
          type: array
          items:
            type: object
            additionalProperties: true
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/ArtifactItem'
        events:
          type: array
          items:
            type: object
            additionalProperties: true
        usage:
          type: object
          additionalProperties: true
        checkpoint:
          type: object
          additionalProperties: true
        checkpoint_id:
          type: string
        task_id:
          type: string
        seed:
          type: integer
        created_at:
          type: string
        updated_at:
          type: string
        parent_rollout_id:
          type: string
        parent_checkpoint_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
        source:
          type: object
          additionalProperties: true
      additionalProperties: true
...
    ValidationError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: array
          items:
            type: string
      required: [code, message]
      additionalProperties: true